Plan code refactor

+++ src/main/java/ea/badge/domain/Plan.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class Plan {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "PlanName", nullable = false)
     private String planName;
@@ -40,7 +40,7 @@ public class Plan {
     private boolean active = true;
 
 
-    public Plan(Integer id, String planName, String planDescription, boolean active) {
+    public Plan(Long id, String planName, String planDescription, boolean active) {
         this.id = id;
         this.planName = planName;
         this.planDescription =planDescription;
